feat: add `DoesNotHaveFile` and `DoesNotHaveDirectory`
Add expectations, that a file or directory does not exist.
