+++ Directory.Packages.props
@@ -3,8 +3,8 @@
 		<ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
 	</PropertyGroup>
 	<ItemGroup>
-		<PackageVersion Include="aweXpect" Version="2.18.0"/>
-		<PackageVersion Include="aweXpect.Core" Version="2.13.0"/>
+		<PackageVersion Include="aweXpect" Version="2.19.1"/>
+		<PackageVersion Include="aweXpect.Core" Version="2.14.0"/>
 		<PackageVersion Include="Testably.Abstractions.Interface" Version="9.0.0"/>
 		<PackageVersion Include="Testably.Abstractions.Testing" Version="4.0.0"/>
 	</ItemGroup>
