+++ Tests/aweXpect.Testably.Tests/HasFile.WithContent.NotSameAs.Tests.cs
@@ -3,7 +3,7 @@ using Testably.Abstractions.Testing;
 
 namespace aweXpect.Testably.Tests;
 
-public partial class HasFile
+public sealed partial class HasFile
 {
 	public sealed partial class WithContent
 	{
