+++ Tests/aweXpect.Testably.Tests/HasFile.WithContent.Tests.cs
@@ -1,9 +1,10 @@
 ﻿using System.IO.Abstractions;
+using System.Text;
 using Testably.Abstractions.Testing;
 
 namespace aweXpect.Testably.Tests;
 
-public partial class HasFile
+public sealed partial class HasFile
 {
 	public sealed partial class WithContent
 	{
@@ -43,6 +44,45 @@ public partial class HasFile
 				IFileSystem sut = new MockFileSystem();
 				// ReSharper disable once MethodHasAsyncOverload
 				sut.File.WriteAllText(path, content);
+
+				async Task Act()
+					=> await That(sut).HasFile(path).WithContent(content);
+
+				await That(Act).DoesNotThrow();
+			}
+		}
+
+		public sealed class BinaryTests
+		{
+			[Fact]
+			public async Task WhenContentIsDifferent_ShouldFail()
+			{
+				byte[] content = Encoding.UTF8.GetBytes("baz");
+				byte[] expected = Encoding.UTF8.GetBytes("bar");
+				string path = "foo.txt";
+				IFileSystem sut = new MockFileSystem();
+				// ReSharper disable once MethodHasAsyncOverload
+				sut.File.WriteAllBytes(path, content);
+
+				async Task Act()
+					=> await That(sut).HasFile(path).WithContent(expected);
+
+				await That(Act).ThrowsException()
+					.WithMessage($"""
+					              Expected that sut
+					              has file '{path}' with content equal to expected,
+					              but it differed
+					              """);
+			}
+
+			[Fact]
+			public async Task WhenContentMatches_ShouldSucceed()
+			{
+				byte[] content = Encoding.UTF8.GetBytes("baz");
+				string path = "foo.txt";
+				IFileSystem sut = new MockFileSystem();
+				// ReSharper disable once MethodHasAsyncOverload
+				sut.File.WriteAllBytes(path, content);
 
 				async Task Act()
 					=> await That(sut).HasFile(path).WithContent(content);
